Lists, Functions, and Classes
This code is designed to take in information about a car such as year, model, etc. It makes a superclass called Vehicle and a sub-class or child class called Automobile and uses those classes to collect the information. After that, it goes to the formatting part, making it print in an easy-to-read format. The cool thing about this code is that the string car could be changed to a variable like V to make the code accept more than the car string.